Move make_sequence_example to magenta.music
