+++ magenta/common/__init__.py
@@ -21,5 +21,4 @@ from .nade import Nade
 from .sequence_example_lib import count_records
 from .sequence_example_lib import flatten_maybe_padded_sequences
 from .sequence_example_lib import get_padded_batch
-from .sequence_example_lib import make_sequence_example
 from .tf_utils import merge_hparams
