+++ magenta/music/melody_encoder_decoder_test.py
@@ -18,7 +18,6 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from magenta.common import sequence_example_lib
 from magenta.music import constants
 from magenta.music import encoder_decoder
 from magenta.music import melodies_lib
@@ -108,7 +107,7 @@ class MelodyOneHotEventSequenceEncoderDecoderTest(tf.test.TestCase):
         [0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
         [0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]]
     expected_labels = [2, 9, 13, 0, 13, 2, 1, 0]
-    expected_sequence_example = sequence_example_lib.make_sequence_example(
+    expected_sequence_example = encoder_decoder.make_sequence_example(
         expected_inputs, expected_labels)
     self.assertEqual(sequence_example, expected_sequence_example)
